+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * Upgraded the CCL parser to a newer and more efficient version. This change will yield general performance improvements in methods that parse CCL statements during evaluation.
 * The test Concourse instance used in a `ClientServerTest` will no longer be automatically deleted when the test fails. This will allow for manual inspection of the instance when debugging the test failure.
 * Fixed a bug that caused the server to fail to start if the `conf/stopwords.txt` configuration file did not exist.
+* Added the ability for the storage engine to track stats and metadata about database structures.
 
 #### Version 0.8.1 (TBD)
 * Fixed a bug that caused local CCL resolution to not work in the `findOrInsert` methods.